refactor(renderers): Move to folder called options instead
