+++ src/options/PlainTextRenderer.js
