+++ src/options/StringRenderer.js
